requirements.txt
annotated-types==0.7.0
anthropic==0.40.0
anyio==3.7.1
appnope==0.1.4
asttokens==3.0.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.3
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.104.1
filelock==3.17.0
Flask==3.0.0
Flask-Cors==4.0.0
fonttools==4.56.0
fsspec==2025.2.0
google-api-core==2.24.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-vision==3.9.0
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
gspread==5.12.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.7.1
httpx==0.28.1
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.1.0
ipykernel==6.30.1
ipython==8.37.0
itsdangerous==2.1.2
jedi==0.19.2
Jinja2==3.1.2
jiter==0.11.0
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
MarkupSafe==2.1.3
matplotlib==3.10.0
matplotlib-inline==0.1.7
MouseInfo==0.1.3
mpmath==1.3.0
mss==10.0.0
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.24.4
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.11.0.86
packaging==24.2
pandas==2.0.3
parso==0.8.5
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
proto-plus==1.25.0
protobuf==5.29.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pydantic==2.5.0
pydantic_core==2.14.1
PyGetWindow==0.0.9
Pygments==2.19.2
pymongo==4.10.1
PyMsgBox==1.0.9
pynput==1.7.7
pyobjc-core==11.0
pyobjc-framework-ApplicationServices==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-CoreText==11.0
pyobjc-framework-Quartz==11.0
pyparsing==3.1.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySide6==6.8.2.1
PySide6_Addons==6.8.2.1
PySide6_Essentials==6.8.2.1
pytest==8.3.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytweening==1.2.0
pytz==2023.3.post1
PyYAML==6.0.2
pyzmq==27.0.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
rubicon-objc==0.5.0
scikit-learn==1.7.1
scipy==1.15.2
seaborn==0.13.2
shiboken6==6.8.2.1
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.27.0
sympy==1.13.1
threadpoolctl==3.6.0
tomli==2.2.1
torch==2.6.0
torchvision==0.21.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2023.4
ultralytics==8.3.78
ultralytics-thop==2.0.14
uritemplate==4.2.0
urllib3==2.1.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.0.1
zipp==3.17.0
